fix(ci): move the ldb drift test to a sibling — I broke the 500 cap again
`gemm_dense.rs` went to 507 when I appended the test inline. That is the same
required check I fixed an hour ago, broken the same way, in the same session.

Moved to `gemm_dense_tests.rs` via the house `#[path]` idiom (417 + 100).

★ The lesson is not "remember the cap" — it is that appending to a file whose
size I have not checked is a coin flip, and I have now lost it twice. Checking
`wc -l` on any file I append to costs nothing; discovering it from a red
required check costs a push cycle.

314 tests green; fmt clean; clippy 0.
